feat(organization): setup of gh org repos, teams and users
This commit adds the core functionality of the IaC management of
repositories, teams and users given by business constraints.

The repository defines the infrastructure live picture within the
`live` folder.

It makes use of 3 modules to handle the different entities that need to
be managed:
- repositories
- teams
- users

There are extra configuration parameters that allow the automatic
generation of Terraform documentation for the project and modules